move health info to separate file
