+++ pkg/admin/prerun/prerun.go
@@ -3,7 +3,6 @@ package prerun
 import (
 	"bytes"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"os"
 	"os/exec"
@@ -14,25 +13,6 @@ import (
 	"github.com/openshift/microshift/pkg/util"
 	"k8s.io/klog/v2"
 )
-
-var (
-	healthFilepath            = "/var/lib/microshift-backups/health.json"
-	errHealthFileDoesNotExist = errors.New("health file does not exist")
-)
-
-type HealthInfo struct {
-	Health       string `json:"health"`
-	DeploymentID string `json:"deployment_id"`
-	BootID       string `json:"boot_id"`
-}
-
-func (hi *HealthInfo) BackupName() data.BackupName {
-	return data.BackupName(fmt.Sprintf("%s_%s", hi.DeploymentID, hi.BootID))
-}
-
-func (hi *HealthInfo) IsHealthy() bool {
-	return hi.Health == "healthy"
-}
 
 type PreRun struct {
 	dataManager data.Manager
@@ -324,25 +304,6 @@ func getCurrentBootID() (string, error) {
 		return "", fmt.Errorf("failed to determine boot ID from %q: %w", path, err)
 	}
 	return strings.ReplaceAll(strings.TrimSpace(string(content)), "-", ""), nil
-}
-
-func getHealthInfo() (*HealthInfo, error) {
-	if exists, err := util.PathExistsAndIsNotEmpty(healthFilepath); err != nil {
-		return nil, err
-	} else if !exists {
-		return nil, errHealthFileDoesNotExist
-	}
-
-	content, err := os.ReadFile(healthFilepath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read health data from %q: %w", healthFilepath, err)
-	}
-
-	health := &HealthInfo{}
-	if err := json.Unmarshal(content, &health); err != nil {
-		return nil, fmt.Errorf("failed to parse health data %q: %w", strings.TrimSpace(string(content)), err)
-	}
-	return health, nil
 }
 
 func getExistingBackupsForTheDeployment(existingBackups []data.BackupName, deployID string) []data.BackupName {
